Bump nelmio/alice to ~2.0
